requirements.txt
asgiref==3.2.3
Django==3.0.3
django-admin==2.0.0
django-excel-response2==2.0.8
django-fontawesome==1.0
django-fontawesome-5==1.0.17
django-geojson==2.12.0
django-leaflet==0.26.0
django-six==1.0.4
djangorestframework==3.11.0
djangorestframework-datatables==0.5.1
geojson==2.5.0
psycopg2==2.8.4
pytz==2019.3
PyYAML==5.3
screen==1.0.1
six==1.14.0
sqlparse==0.3.0
xlwt==1.3.0
